Add support for dynamic board sizes
This commit introduces a command line flag `-s` to specify the board size, which must be a positive integer. The default size is 9 (3x3 board).

Additionally, a `deque` data structure is now used for the open list of nodes, improving the speed of pop and push operations. Note that this feature is currently disabled for informed search algorithms, as their frontier is implemented as a priority queue.